Fix an encoding bug for list of tuples.
If an argument is of type "tuple" it's encoded as the concatenation of
the types that are inside. We should do the same for list of tuples.

This fixes a bug where brownie crashes when trying to encode an argument
with a type "ComplexStruct[]" where ComplexStruct is a solidity struct
with a few fields.

Example Solidity code to reproduce the bug:

contract Foo {
    struct Bar {
        uint256 something;
        address something_else;
    }

    function frobulate(Bar[] calldata bar_list) external {
      // do something
    }
}

Before that fix, Brownie will crash when encoding the call to frobulate.