Add note about where to run `brownie init`
I just ran `brownie init` in my `~/src` directory thinking it would prompt for the name of the project and it ended up making the files where I ran the command. We should add a note that `init` should be run from inside the new project's directory.

I'm not sure the best way to word it though.