fix: handle all calls are queried from the same block
Previously, if the block_identifier was not specified and the pending
calls queue was flushed, any future pending calls would come from a
different block. To handle this, we now on init hardcode the block
identifier to call from if not specified. Also handled is if the block
identifier is specified in a dev network, and the address is not
supplied, the user needs to deploy multicall2 first. (we raise a
ContractNotFound error, since the contract was not found at user
specified block height)