Avoid KeyError exception when setting active contract
When `set_active_contract` was called on a contract that has already been set
as active before, the loop updating the values of `self.pcMap` would raise a
`keyError` exception.
That is because the value of the `path` keys were already updated with their
corresponding path taken from the map in `self.pathMap`. The updated paths
are the values of `self.pathMap` not the keys, therefore upon any subsequent
lookup attempt to retrieve them from `self.pathMap` a `keyError` exception
will be raised.